+++ src/main/java/com/github/wz2cool/dynamic/mybatis/ColumnInfo.java
@@ -11,6 +11,9 @@ class ColumnInfo {
     private String columnName;
     private String tableOrAlias;
     private boolean updateIfNull;
+    private boolean insertIfNull;
+    private boolean insertIgnore;
+    private JdbcType jdbcType;
     private Field field;
 
     public Field getField() {
@@ -51,5 +54,29 @@ class ColumnInfo {
         } else {
             return this.columnName;
         }
+    }
+
+    public boolean isInsertIfNull() {
+        return insertIfNull;
+    }
+
+    public void setInsertIfNull(boolean insertIfNull) {
+        this.insertIfNull = insertIfNull;
+    }
+
+    public JdbcType getJdbcType() {
+        return jdbcType;
+    }
+
+    public void setJdbcType(JdbcType jdbcType) {
+        this.jdbcType = jdbcType;
+    }
+
+    public boolean isInsertIgnore() {
+        return insertIgnore;
+    }
+
+    public void setInsertIgnore(boolean insertIgnore) {
+        this.insertIgnore = insertIgnore;
     }
 }