+++ src/main/java/com/github/wz2cool/dynamic/mybatis/EntityHelper.java
@@ -56,12 +56,38 @@ class EntityHelper {
         return camelCaseToDashCase(usePropertyName);
     }
 
+    static JdbcType getJdbcTypeByProperty(final String propertyName, final Field[] properties) {
+        Column column = getColumnByProperty(propertyName, properties);
+        if (column == null) {
+            return null;
+        }
+        return column.jdbcType();
+    }
+
     static boolean isPropertyUpdateIfNull(final String propertyName, final Field[] properties) {
+        Column column = getColumnByProperty(propertyName, properties);
+        if (column == null) {
+            // default don't update property.
+            return false;
+        }
+        return column.updateIfNull();
+    }
+
+    static boolean isPropertyInsertIfNull(final String propertyName, final Field[] properties) {
+        Column column = getColumnByProperty(propertyName, properties);
+        if (column == null) {
+            // default insert property.
+            return false;
+        }
+        return column.insertIfNull();
+    }
+
+    static boolean isPropertyInsertIgnore(final String propertyName, final Field[] properties) {
         Column column = getColumnByProperty(propertyName, properties);
         if (column == null) {
             return false;
         }
-        return column.updateIfNull();
+        return column.insertIgnore();
     }
 
     static Column getColumnByProperty(final String propertyName, final Field[] properties) {