+++ src/main/java/com/github/wz2cool/helper/CommonsHelper.java
@@ -1,8 +1,11 @@
 package com.github.wz2cool.helper;
 
+import jodd.methref.Methref;
+
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.function.Function;
 
 /**
  * Created by Frank on 7/7/2017.
@@ -64,5 +67,20 @@ public class CommonsHelper {
         }
 
         return obj.toString();
+    }
+
+    public static <T> String getPropertyName(final Class<T> target, final Function<T, Object> getMethodFunc) {
+        String methodName = obtainGetMethodName(target, getMethodFunc);
+        if (methodName.startsWith("get")) {
+            return java.beans.Introspector.decapitalize(methodName.substring(3, methodName.length()));
+        }
+
+        return methodName;
+    }
+
+    public static synchronized <T> String obtainGetMethodName(final Class<T> target, final Function<T, Object> getMethodFunc) {
+        Methref<T> methodRef = Methref.on(target);
+        getMethodFunc.apply(methodRef.to());
+        return methodRef.ref();
     }
 }